restructure collectors directory
This commit restructures the collector/ directory to collectors/ in
order to remain consistent with the pattern we've already established
with the producers/ directory.

It also moves the HTTP client out of the collectors/ directory and into
the util/http/client/ directory, which better matches its purpose.

And best of all: all of our package names aren't mixedCase NOR contain
underscores! :tada:
