+++ collectors/framework/record.go
