+++ collectors/framework/record_test.go
