+++ collectors/interface.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package collector
+package collectors
 
 // NodeInfo represents information about the node, such as the IP address,
 // hostname, Mesos ID, and the cluster ID that the node belongs to.
